backend/tax/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pandas==2.2.3
openpyxl==3.1.5
google-genai==1.0.0
anthropic
groq
python-dotenv==1.0.1
python-multipart==0.0.12

# SQLAlchemy ORM + DB
sqlalchemy==2.0.36

# Authentication — JWT + password hashing
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# PostgreSQL driver
psycopg2-binary==2.9.9

# Redis client
redis==5.0.8
